# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для телефонного справочника",
        "title": "Telephone Book API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход пользователя",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/login.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/register.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/register.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/departments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "Получить все отделы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/departments.DepartmentsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "Обновить отдел",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Старое название отдела",
                        "name": "department",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Новые данные отдела",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/departments.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/departments.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "Создать отдел",
                "parameters": [
                    {
                        "description": "Данные отдела",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/departments.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/departments.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "Удалить отдел",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название отдела",
                        "name": "department",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/departments.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "departments.CreateRequest": {
            "type": "object",
            "properties": {
                "institute": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sections": {
                    "description": "Optional, can be used to specify sections within the department",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "departments.CreateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                },
                "user_id": {
                    "description": "ID созданного отдела",
                    "type": "integer"
                }
            }
        },
        "departments.DeleteResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
        "departments.DepartmentsResponse": {
            "type": "object",
            "properties": {
                "departments": {
                    "description": "Список отделов",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Department"
                    }
                },
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
...
        "departments.UpdateRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "sections": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "departments.UpdateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
...
        "login.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "login.LoginResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Response базовая структура ответа\nСтатус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                },
                "token": {
                    "description": "JWT токен. Пример: \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
                    "type": "string"
                }
            }
        },
        "models.Department": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "register.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "register.RegisterResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "description": "Базовый ответ API",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
...
    }
}